license!: switched from boost software license v1 to mozilla public license v2
This change was motivated by the problem of having to have a CLA for bsl-1 to be viable.
I'd have to ask of all future contributors to accept the bsl-1 license.
MPL-2.0 does this more or less automatic.

Moreover philosphy...
I am torn between permissive and copyleft, so I ended up here.