fix: unexpected bad performance during build script and benchmark with simultanious cargo call
There seems to exist some kind of locking or overhead, when calling cargo in the .rustup cargo folder,
compared to cargo to the main toolchain. This fix just adds as default the parsing of the CARGO env var
which in turn makes this library use the cargo implementation of the toolchain this lib is compiled with.
On the other hand this brings the bench results inline with the results from v0.7.3, which also used
the cargo from the default toolchain.